+++ app/src/main/java/com/battlelancer/seriesguide/movies/TmdbMoviesDataSource.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2020-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.movies
 
@@ -25,7 +25,7 @@ import java.util.Calendar
 import java.util.Date
 
 /**
- * Loads movies from TMDb in chunks.
+ * Loads movies from TMDb in pages.
  *
  * If a query is given, will load search results for that query. Otherwise will load a list of
  * movies based on the given link.
@@ -37,6 +37,8 @@ class TmdbMoviesDataSource(
     private val query: String,
     private val languageCode: String,
     private val regionCode: String,
+    private val releaseYear: Int?,
+    private val originalLanguageCode: String?,
     private val watchProviderIds: List<Int>?,
     private val watchRegion: String?
 ) : PagingSource<Int, BaseMovie>() {
@@ -76,11 +78,13 @@ class TmdbMoviesDataSource(
         val action: String
         val call: Call<MovieResultsPage>
         if (query.isEmpty()) {
-            val pair = buildMovieListCall(languageCode, regionCode, pageNumber)
+            val pair = buildMovieListCall(pageNumber)
             action = pair.first
             call = pair.second
         } else {
             action = "search for movies"
+            // Use year instead of primary_release_year as movies may get released years apart
+            // in different regions.
             call = tmdb.searchService()
                 .movie(
                     query,
@@ -88,7 +92,7 @@ class TmdbMoviesDataSource(
                     languageCode,
                     regionCode,
                     false,
-                    null,
+                    releaseYear,
                     null
                 )
         }
@@ -139,17 +143,21 @@ class TmdbMoviesDataSource(
         }
     }
 
-    private fun buildMovieListCall(
-        languageCode: String?,
-        regionCode: String?,
-        page: Int
-    ): Pair<String, Call<MovieResultsPage>> {
+    private fun buildMovieListCall(page: Int): Pair<String, Call<MovieResultsPage>> {
         val builder = tmdb.discoverMovie()
             .language(languageCode)
             .region(regionCode)
             .page(page)
+        if (supportsYearFilter(link) && releaseYear != null) {
+            // Use year instead of primary_release_year as movies may get released years apart
+            // in different regions.
+            builder.year(releaseYear)
+        }
+        if (originalLanguageCode != null) {
+            builder.with_original_language(originalLanguageCode)
+        }
         // Only filter by watch provider if release type DIGITAL included.
-        if (isLinkFilterable(link)) {
+        if (supportsWatchProviderFilter(link)) {
             if (!watchProviderIds.isNullOrEmpty() && watchRegion != null) {
                 builder
                     .with_watch_providers(DiscoverFilter(OR, *watchProviderIds.toTypedArray()))
@@ -219,8 +227,11 @@ class TmdbMoviesDataSource(
     }
 
     companion object {
-        fun isLinkFilterable(link: MoviesDiscoverLink): Boolean =
+        fun supportsWatchProviderFilter(link: MoviesDiscoverLink): Boolean =
             link == MoviesDiscoverLink.POPULAR || link == MoviesDiscoverLink.DIGITAL
+
+        fun supportsYearFilter(link: MoviesDiscoverLink): Boolean =
+            link == MoviesDiscoverLink.POPULAR
     }
 
 }